+++ packages/desktop/src/main/main.ts
@@ -319,7 +319,11 @@ const closeBackendProcess = (clear: boolean = false) => {
       if (clear) {
         resetting = false
         const appData = app.getPath('appData')
-        fs.rmSync(appData, { recursive: true, force: true })
+        try {
+          fs.rmSync(appData, { recursive: true, force: true })
+        } catch (e) {
+          console.error(e)
+        }
       }
       app.quit()
     }, 2000)
@@ -331,7 +335,11 @@ const closeBackendProcess = (clear: boolean = false) => {
         if (clear) {
           resetting = false
           const appData = app.getPath('appData')
-          fs.rmSync(appData, { recursive: true, force: true })
+          try {
+            fs.rmSync(appData, { recursive: true, force: true })
+          } catch (e) {
+            console.error(e)
+          }
         }
         app.quit()
       }