+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/TryQuiet/quiet/compare/quiet@1.1.0-alpha.6...quiet@1.1.0) (2023-03-20)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.1.0-alpha.6](https://github.com/TryQuiet/quiet/compare/quiet@1.1.0-alpha.5...quiet@1.1.0-alpha.6) (2023-03-17)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.1.0-alpha.5](https://github.com/TryQuiet/quiet/compare/quiet@1.1.0-alpha.4...quiet@1.1.0-alpha.5) (2023-03-17)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.1.0-alpha.4](https://github.com/TryQuiet/quiet/compare/quiet@1.1.0-alpha.3...quiet@1.1.0-alpha.4) (2023-03-15)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
 # [1.1.0-alpha.3](https://github.com/TryQuiet/quiet/compare/quiet@1.1.0-alpha.2...quiet@1.1.0-alpha.3) (2023-03-13)
 
 **Note:** Version bump only for package quiet